Move server init to its own package
